+++ packages/headless/src/features/insight-search/insight-search-actions.ts
@@ -1,11 +1,17 @@
 import {createAsyncThunk} from '../..';
+import {
+  historyStore,
+  StateNeededByInsightAnalyticsProvider,
+} from '../../api/analytics/insight-analytics';
 import {isErrorResponse} from '../../api/search/search-api-client';
+import {SearchResponseSuccess} from '../../api/search/search/search-response';
 import {
   AsyncThunkInsightOptions,
   InsightAPIClient,
 } from '../../api/service/insight/insight-api-client';
 import {InsightQueryRequest} from '../../api/service/insight/query/query-request';
 import {
+  CategoryFacetSection,
   ConfigurationSection,
   FacetSection,
   InsightCaseContextSection,
@@ -15,11 +21,8 @@ import {
   SearchSection,
 } from '../../state/state-sections';
 import {requiredNonEmptyString} from '../../utils/validate-payload';
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-  SearchAction,
-} from '../analytics/analytics-utils';
+import {InsightAction} from '../analytics/analytics-utils';
+import {CategoryFacetSetState} from '../facets/category-facet-set/category-facet-set-state';
 import {AnyFacetRequest} from '../facets/generic/interfaces/generic-facet-request';
 import {snapshot} from '../history/history-actions';
 import {extractHistory} from '../history/history-state';
@@ -29,8 +32,13 @@ import {
   FetchQuerySuggestionsThunkReturn,
   StateNeededByQuerySuggest,
 } from '../query-suggest/query-suggest-actions';
+import {getQueryInitialState} from '../query/query-state';
 import {ExecuteSearchThunkReturn} from '../search/search-actions';
-import {logQueryError} from '../search/search-analytics-actions';
+import {getSearchInitialState} from '../search/search-state';
+import {
+  logFetchMoreResults,
+  logQueryError,
+} from './insight-search-analytics-actions';
 
 export type StateNeededByExecuteSearch = ConfigurationSection &
   InsightConfigurationSection &
@@ -39,6 +47,7 @@ export type StateNeededByExecuteSearch = ConfigurationSection &
       SearchSection &
       QuerySection &
       FacetSection &
+      CategoryFacetSection &
       PaginationSection
   >;
 
@@ -56,16 +65,16 @@ const fetchFromAPI = async (
 
 export const executeSearch = createAsyncThunk<
   ExecuteSearchThunkReturn,
-  SearchAction,
+  InsightAction,
   AsyncThunkInsightOptions<StateNeededByExecuteSearch>
 >(
   'search/executeSearch',
   async (
-    analyticsAction: SearchAction,
+    analyticsAction: InsightAction,
     {getState, dispatch, rejectWithValue, extra}
   ) => {
-    /** TODO: We need to dispatch analytics action, but first we have to create InsightClientProvider so the refactor will be available in  https://coveord.atlassian.net/browse/SVCC-2246*/
     const state = getState();
+    addEntryInActionsHistory(state);
     const fetched = await fetchFromAPI(
       extra.apiClient,
       state,
@@ -77,6 +86,18 @@ export const executeSearch = createAsyncThunk<
       return rejectWithValue(fetched.response.error);
     }
 
+    const fetchedResponse = fetched.response.success;
+    analyticsAction(
+      dispatch,
+      () =>
+        getStateAfterResponse(
+          fetched.queryExecuted,
+          fetched.duration,
+          state,
+          fetchedResponse
+        ),
+      extra
+    );
     dispatch(snapshot(extractHistory(getState())));
 
     return {
@@ -122,12 +143,12 @@ export const fetchMoreResults = createAsyncThunk<
 
 export const fetchFacetValues = createAsyncThunk<
   ExecuteSearchThunkReturn,
-  SearchAction,
+  InsightAction,
   AsyncThunkInsightOptions<StateNeededByExecuteSearch>
 >(
   'search/fetchFacetValues',
   async (
-    analyticsAction: SearchAction,
+    analyticsAction: InsightAction,
     {getState, dispatch, rejectWithValue, extra: {apiClient}}
   ) => {
     const state = getState();
@@ -182,13 +203,14 @@ export const fetchQuerySuggestions = createAsyncThunk<
 const buildInsightSearchRequest = (
   state: StateNeededByExecuteSearch
 ): InsightQueryRequest => {
+  const facets = getAllFacets(state);
   return {
     accessToken: state.configuration.accessToken,
     organizationId: state.configuration.organizationId,
     url: state.configuration.platformUrl,
     insightId: state.insightConfiguration.insightId,
     q: state.query?.q,
-    facets: getFacetRequests(state.facetSet),
+    ...(facets.length && {facets}),
     caseContext: state.insightCaseContext?.caseContext,
     ...(state.pagination && {
       firstResult: state.pagination.firstResult,
@@ -217,12 +239,16 @@ const buildInsightFetchFacetValuesRequest = (
   };
 };
 
-export const logFetchMoreResults = makeAnalyticsAction(
-  'search/logFetchMoreResults',
-  AnalyticsType.Search,
-  (client) => client.logFetchMoreResults()
-);
+function getAllFacets(state: StateNeededByExecuteSearch) {
+  return [
+    ...getFacetRequests(state.facetSet),
+    ...getCategoryFacetRequests(state.categoryFacetSet),
+  ];
+}
 
+function getCategoryFacetRequests(state: CategoryFacetSetState | undefined) {
+  return Object.values(state || {}).map((slice) => slice!.request);
+}
 function getFacetRequests<T extends AnyFacetRequest>(
   requests: Record<string, T> = {}
 ) {
@@ -231,3 +257,41 @@ function getFacetRequests<T extends AnyFacetRequest>(
 
 const getOriginalQuery = (state: StateNeededByExecuteSearch) =>
   state.query?.q !== undefined ? state.query.q : '';
+
+const getStateAfterResponse: (
+  query: string,
+  duration: number,
+  previousState: StateNeededByExecuteSearch,
+  response: SearchResponseSuccess
+) => StateNeededByInsightAnalyticsProvider = (
+  query,
+  duration,
+  previousState,
+  response
+) => ({
+  ...previousState,
+  query: {
+    q: query,
+    enableQuerySyntax:
+      previousState.query?.enableQuerySyntax ??
+      getQueryInitialState().enableQuerySyntax,
+  },
+  search: {
+    ...getSearchInitialState(),
+    duration,
+    response,
+    results: response.results,
+  },
+});
+
+const addEntryInActionsHistory = (state: StateNeededByExecuteSearch) => {
+  if (state.configuration.analytics.enabled) {
+    historyStore.addElement({
+      name: 'Query',
+      ...(state.query?.q && {
+        value: state.query.q,
+      }),
+      time: JSON.stringify(new Date()),
+    });
+  }
+};