+++ packages/bueno/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.41.0](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.40.9...@coveo/bueno@0.41.0) (2022-07-06)
+
+
+### Features
+
+* **headless:** Add analytics actions to insight-search ([#2147](https://github.com/coveo/ui-kit/issues/2147)) ([e5c72ad](https://github.com/coveo/ui-kit/commit/e5c72ad5c2073dd7ebeb3afc087246735c99a4e2))
+
+
+
+
+
+## [0.40.9](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.40.8...@coveo/bueno@0.40.9) (2022-07-05)
+
+
+### Bug Fixes
+
+* **atomic:** add explicit types for different types of store ([#2190](https://github.com/coveo/ui-kit/issues/2190)) ([16afb1e](https://github.com/coveo/ui-kit/commit/16afb1ea91a4b97ca5a34585b342dc6208dc9ed0))
+
+
+
+
+
+## [0.40.8](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.40.7...@coveo/bueno@0.40.8) (2022-07-05)
+
+**Note:** Version bump only for package @coveo/bueno
+
+
+
+
+
 ## [0.40.7](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.40.6...@coveo/bueno@0.40.7) (2022-07-05)