+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.27.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.26.9...@coveo/quantic@1.27.0) (2022-07-06)
+
+
+### Features
+
+* **headless:** Add analytics actions to insight-search ([#2147](https://github.com/coveo/ui-kit/issues/2147)) ([e5c72ad](https://github.com/coveo/ui-kit/commit/e5c72ad5c2073dd7ebeb3afc087246735c99a4e2))
+
+
+
+
+
+## [1.26.9](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.26.8...@coveo/quantic@1.26.9) (2022-07-05)
+
+
+### Bug Fixes
+
+* **atomic:** add explicit types for different types of store ([#2190](https://github.com/coveo/ui-kit/issues/2190)) ([16afb1e](https://github.com/coveo/ui-kit/commit/16afb1ea91a4b97ca5a34585b342dc6208dc9ed0))
+
+
+
+
+
+## [1.26.8](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.26.7...@coveo/quantic@1.26.8) (2022-07-05)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [1.26.7](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.26.6...@coveo/quantic@1.26.7) (2022-07-05)